fix(pager) add order details
